backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
filelock==3.18.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.7.0
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain-core==1.2.7
langchain-huggingface==1.2.0
langsmith==0.6.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==2.2.6
openai==2.15.0
opencv-python==4.12.0.88
orjson==3.11.5
ormsgpack==1.12.1
packaging==24.2
pillow==12.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
polars==1.35.2
polars-runtime-32==1.35.2
propcache==0.4.1
psutil==7.1.3
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2026.1.15
requests==2.32.4
requests-toolbelt==1.0.0
scipy==1.16.3
simsimd==6.5.12
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.14.1
ultralytics==8.3.228
ultralytics-thop==2.0.18
urllib3==2.5.0
uuid_utils==0.13.0
uv==0.7.8
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
sentence-transformers
langchain
fastapi
uvicorn[standard]
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain-groq
langchain-community
langchain-experimental
langchain-core
jinja2
pydantic
-e .
